requirements.txt
audioread==2.1.8
cffi==1.14.0
decorator==4.4.2
joblib==0.15.1
librosa==0.6.1
llvmlite==0.31.0
numba==0.48.0
numpy==1.19.0
Pillow==7.2.0
protobuf==3.12.2
pycparser==2.20
resampy==0.2.2
scikit-learn==0.23.1
scipy==1.5.0
six==1.15.0
SoundFile==0.10.3.post1
tensorboardX==2.0
threadpoolctl==2.1.0
torch==1.4.0+cu100
torchvision==0.5.0+cu100
